credentials: retry calls on failures
Add a retry mechanism for the session
created in get_session() and use that
inside the Credentials object.

The Retry class from urllib3 respects
the Retry-After header, so that should
hopefully avoid these types of failures
in the future:

HTTPError: 503 Server Error: Service Unavailable for url:
https://identity.dataspace.copernicus.eu/auth/realms/CDSE/.well-known/openid-configuration

which I saw raised inside __openid_configuration().