download: avoid sleep on exit
Hoist the sleep to the first thing
in the loop so the final failure
immediately returns instead
of sleeping for 1+ minute.

Also put the increment of attempts
there to avoid duplicating the same
increment before every continue.

This fixes a potential hang when
getting TokenClientConnectionError
repeatedly when getting a session.