download: retry ChunkedEncodingError
The server might not send all bytes specified
by the Content-Length header before closing
the connection. Log a warning and try again
when that happens.

This triggers more exceptions, and one of
them is because some underlying library
is trying to close fd -1, so wrap
most of the loop body in a try to catch
that.